# siekanski-private-lessons-backend/Models/UserRegistratrionModel.cs
namespace siekanski_private_lessons_backend.Models
{
    public class UserRegistrationModel
    {
        public string FirstName { get; set; }
        public string LastName { get; set; }
        public string Username { get; set; }
        public string Email { get; set; }
        public string Password { get; set; }
        public string Role { get; set; } // "Student" or "Teacher"
    }
}